move enum NewState to states package and delete unused tiledMapRenderer
