+++ android/src/car/superfun/game/googleGamePlayServices/SetUpGame.java
@@ -29,7 +29,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import car.superfun.game.AndroidLauncher;
-import car.superfun.game.NewState;
+import car.superfun.game.states.NewState;
 import car.superfun.game.R;
 import car.superfun.game.car.OpponentCarController;
